Moved project tasks
